(REF) separate atlas mixins from main common atlas classes
